+++ test/transaction.js
@@ -1,77 +1,83 @@
 var should = require('should');
 
 var grex = require('../');
-var client = grex.createClient();
+var bind = grex.bindParameter;
 var g = grex.g;
 var gremlin = grex.gremlin;
 
-var alice, bob;
-var james, waldo;
 
-describe('Transaction commit', function() {
-  describe('when adding elements to the graph in a transaction', function() {
-    it('should add a vertex in a transaction', function(done) {
-      var query = gremlin(g.addVertex({ name: "Alice" }));
+function cleanGraph(done) {
+  var client = grex.createClient();
+  var query = gremlin();
 
-      client.execute(query, function(err, result) {
-        should.not.exist(err);
-        result.should.have.property('success', true);
-        done();
-      });
+  query(g.V('name', 'Alice').remove());
+  query(g.V('name', 'Bob').remove());
+  query(g.V('name', 'Carol').remove());
+  query(g.V('name', 'Eve').remove());
+  query(g.V('name', 'John').remove());
+  query(g.V('name', 'Waldo').remove());
+
+  client.execute(query, function(err, results) {
+    done();
+  });
+}
+
+describe('Transactions', function() {
+  afterEach(cleanGraph);
+
+  it('should add a vertex to the graph', function(done) {
+    var client = grex.createClient();
+    var query = gremlin(g.addVertex({ name: "Alice" }));
+
+    client.execute(query, function(err, result) {
+      should.not.exist(err);
+      result.should.have.property('success', true);
+      done();
     });
-
-    it('should add two vertices and an edge in a transaction', function(done) {
-      var query = gremlin();
-      bob = query.var(g.addVertex({ name: 'Bob' }));
-      waldo = query.var(g.addVertex({ name: 'Waldo' }));
-      query(g.addEdge(bob, waldo, 'likes', { since: 'now' }));
-
-      query.script.split('\n').length.should.equal(4);
-      client.execute(query, function(err, result) {
-        should.not.exist(err);
-        result.should.have.property('success', true);
-        done();
-      });
-    });
-
-    // Clean up: remove james and waldo from the database
-    // after(function(done) {
-    //   var gremlin = grex.gremlin;
-    //   james.remove();
-    //   waldo.remove();
-
-    //   gremlin.execute()
-    //   .then(function(){
-    //     done();
-    //   });
-    // });
   });
 
-  describe('when deleting two vertices', function() {
-    before(function(done) {
-      g.V('name', 'Jess').get(function(err, result) {
-        alice = result.results[0];
-        done();
-      });
+  it('should add a vertex with bound properties to the graph', function(done) {
+    var client = grex.createClient();
+    var query = gremlin(g.addVertex(bind({ name: "Eve" })));
+
+    query.script.should.equal('g.addVertex(p0)\n');
+    query.params.p0.name.should.equal('Eve');
+
+    client.execute(query, function(err, result) {
+      should.not.exist(err);
+      result.should.have.property('success', true);
+      done();
     });
+  });
 
-    before(function(done) {
-      g.V('name', 'Bob').get(function(err, result) {
-        bob = result.results[0];
-        done();
-      });
+  it('should add two vertices and an edge in a transaction', function(done) {
+    var client = grex.createClient();
+    var query = gremlin();
+    var bob = query.var(g.addVertex({ name: 'Bob' }));
+    var waldo = query.var(g.addVertex({ name: 'Waldo' }));
+
+    query(g.addEdge(bob, waldo, 'likes', { since: 'now' }));
+
+    query.script.split('\n').length.should.equal(4);
+    client.execute(query, function(err, result) {
+      should.not.exist(err);
+      result.should.have.property('success', true);
+      done();
     });
+  });
 
-    // it('should remove vertices in a transaction', function(done) {
-    //   var gremlin = grex.gremlin;
+  it('should add two vertices and an edge with bound properties', function(done) {
+    var client = grex.createClient();
+    var query = gremlin();
+    var john = query.var(g.addVertex(bind({ name: 'John' })));
+    var carol = query.var(g.addVertex(bind({ name: 'Carol' })));
+    query(g.addEdge(john, carol, 'knows', bind({ since: 'now' })));
 
-    //   alice.remove();
-    //   bob.remove();
-
-    //   gremlin.execute()
-    //   .then(function(result) {
-    //     done();
-    //   });
-    // });
+    query.script.split('\n').length.should.equal(4);
+    client.execute(query, function(err, result) {
+      should.not.exist(err);
+      result.should.have.property('success', true);
+      done();
+    });
   });
 });